SoraRaw: Refactor to align with contributing guidelines
